requirements.txt
anthropic
requests
python-dotenv
pandas
pandas-ta
matplotlib
mplfinance
alpaca-trade-api
oandapyV20
pycoingecko
apscheduler
python-telegram-bot
aiohttp
numpy
